+++ src/Entities/Abstracts/Order.php
@@ -23,7 +23,7 @@ abstract class Order implements iOrder
     const TYPE_BUY = 1;
     const TYPE_SELL = 2;
 
-    public function __construct(iPair $pair, int $quantity, int $price, iBalance $balance, int $orderNumber = null, int $type)
+    public function __construct(iPair $pair, int $quantity, float $price, iBalance $balance, int $orderNumber = null, int $type)
     {
         $this->pair = $pair;
         $this->quantity = $quantity;
@@ -75,7 +75,7 @@ abstract class Order implements iOrder
         return $this->type;
     }
 
-    public function getPrice() : int
+    public function getPrice() : float
     {
         return $this->price;
     }
@@ -90,12 +90,12 @@ abstract class Order implements iOrder
         return $this->quantity_remain;
     }
 
-    public function getCost() : int
+    public function getCost() : float
     {
         return $this->quantity * $this->price;
     }
 
-    public function getCostRemain() : int
+    public function getCostRemain() : float
     {
         return $this->quantity_remain * $this->price;
     }
